requirements.txt
discord.py
pandas
